# infra/power-platform/custom-connector/apiDefinition.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Copilot Agent ROI Tracker",
    "description": "Custom connector for the Copilot Agent ROI Tracker. Enables Power Platform flows to query agent data, record usage events, and view dashboard metrics.",
    "version": "1.0.0",
    "contact": {
      "name": "ROI Tracker Support",
      "url": "https://github.com/your-org/copilot-roi-tracker"
    }
  },
  "host": "your-app.azurewebsites.net",
  "basePath": "/api",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "oauth2_auth": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
      "tokenUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/token",
      "scopes": {}
    }
  },
  "paths": {
...
    "/usage": {
      "post": {
        "operationId": "RecordUsage",
        "summary": "Record Agent Usage",
        "description": "Record a single agent usage event for ROI tracking",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["agentId", "actionBenchmarkId", "userId", "userName"],
              "properties": {
                "agentId": { "type": "string", "description": "Agent ID" },
                "actionBenchmarkId": { "type": "string", "description": "Benchmark action ID" },
                "userId": { "type": "string", "description": "User ID or email" },
                "userName": { "type": "string", "description": "User display name" },
                "notes": { "type": "string", "description": "Optional notes" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usage recorded",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "agentId": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/usage/bulk": {
      "post": {
        "operationId": "BulkRecordUsage",
        "summary": "Bulk Record Usage",
        "description": "Record multiple usage events at once",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "agentId": { "type": "string" },
                  "actionBenchmarkId": { "type": "string" },
                  "userId": { "type": "string" },
                  "userName": { "type": "string" }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usage events recorded"
          }
        }
      }
    },
...
  }
}